e2e_samples/parking_sensors_synapse/tests/integrationtests/requirements.txt
pip
wheel
watchdog
flake8
coverage
Sphinx
twine
pytest
pytest-runner
pytest-adf
tox
autopep8
azure-mgmt-resource
azure-mgmt-datafactory
azure-storage-blob
pyodbc
azure-synapse==0.1.1
azure-synapse-artifacts==0.9.0
azure-storage-file-datalake
